opt/otp-service/requirements.txt
fastapi
uvicorn
pyotp
qrcode
pillow
mysql-connector-python
passlib[bcrypt]
python-multipart
